+++ tests/movement_2d.rs
@@ -5,7 +5,7 @@ use hyperchess::domain::rules::Rules;
 use std::collections::HashSet;
 
 fn coord(x: usize, y: usize) -> Coordinate {
-    Coordinate::new(vec![x, y])
+    Coordinate::new(vec![x as u8, y as u8])
 }
 
 #[test]
